move get_cases_from_forms to processor
